Add multiplication task

+++ multiplication/multiplication.h
@@ -2,6 +2,6 @@
 
 #include <stdexcept>
 
-int64_t Multiply(int a, int b) {
+int64_t Multiply(int64_t a, int64_t b) {
     return a * b;
 }
